+++ iExtra/UI/Extensions/UIView/UIView+Rotate.swift
@@ -10,6 +10,8 @@ import UIKit
 
 public extension UIView {
     
+    private var key: String { return "rotation" }
+    
     public func rotate(
         withRadians radians: CGFloat,
         duration: TimeInterval,
@@ -22,5 +24,32 @@ public extension UIView {
             animations: { self.transform = self.transform.rotated(by: radians) },
             completion: completion
         )
+    }
+    
+    public func rotate360(withDuration duration: Double, times: Int) {
+        let animation = rotate360Animation(withDuration: duration, times: times)
+        layer.add(animation, forKey: key)
+    }
+    
+    public func startRotation(withDuration duration: Double) {
+        rotate360(withDuration: duration, times: Int.max)
+    }
+    
+    public func stopRotation() {
+        layer.removeAnimation(forKey: key)
+    }
+}
+
+private extension UIView {
+    
+    func rotate360Animation(withDuration duration: Double, times: Int) -> CABasicAnimation {
+        let animation = CABasicAnimation(keyPath: "transform.rotation")
+        animation.duration = duration / 2
+        animation.fromValue = 0
+        animation.toValue = CGFloat(Double.pi)
+        animation.timingFunction = CAMediaTimingFunction(name: kCAMediaTimingFunctionLinear)
+        animation.autoreverses = false
+        animation.repeatCount = Float(times)
+        return animation
     }
 }